BetalSkat: Delt klassen op i to, og implementeret dem
