+++ src/main/java/spillogik/SpilGenerator.java
@@ -42,8 +42,7 @@ public class SpilGenerator {
         //-----------------------------------------Variabler--------------------------------------//
         EjendomCO ejendom;
         ChanceFeltCO proevLykken;
-        BetalSkat indkomstSkat;
-        BetalSkat statsSkat;
+        StatsSkat statsSkat;
         Rederi rederi;
         Bryggeri bryggeri;
         FaengselCO faengsel;
@@ -79,7 +78,7 @@ public class SpilGenerator {
         ejendomsGruppe.tilfoejEjendom(ejendom);
         felter[3] = ejendom;
 
-        indkomstSkat = new BetalSkat(4);
+        IndkomstSkat indkomstSkat = new IndkomstSkat(4);
         felter[4] = indkomstSkat;
 
         rederi = new Rederi("Øresund A/S",200,5);
@@ -321,7 +320,7 @@ public class SpilGenerator {
         ejendomsGruppe.tilfoejEjendom(ejendom);
         felter[37] = ejendom;
 
-        statsSkat = new BetalSkat(38);
+        statsSkat = new StatsSkat(38, 100);
         felter[38] = statsSkat;
 
         ejendom = new EjendomCO("Rådhuspladsen",400,50,39);
